testing new root combine script

+++ run_Prod_Analysis.sh
@@ -574,13 +574,20 @@ if [[ $a_flag = "true" ]]; then
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_all_RF" "${OutDATAFilename}_Left" "${data_left[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_prompt_RF" "${OutDATAFilename}_Left" "${data_left[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_rand_RF" "${OutDATAFilename}_Left" "${data_left[*]}"
-	for i in "${data_left[@]}"
-	do
-	    cd "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT"
-	    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
-	    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	done
+	if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${OutDATAFilename}_Left.root" ]; then
+	    for i in "${data_left[@]}"
+	    do
+		if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${i}_-1_Raw_Data.root" ]; then
+		    cd "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT"
+		    echo "Combining run $i with ${OutDATAFilename}_Left.root..."
+		    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
+		    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		else
+		    echo "WARNING: ${i}_Raw_Data.root does not exist!"
+		fi
+	    done	 
+	fi
 	echo
     fi
 
@@ -602,15 +609,22 @@ if [[ $a_flag = "true" ]]; then
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_all_RF" "${OutDATAFilename}_Center" "${data_center[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_prompt_RF" "${OutDATAFilename}_Center" "${data_center[*]}"
 	#python3 mergeRootFiles.py "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/" "_-1_Raw_Data" "Cut_Kaon_Events_rand_RF" "${OutDATAFilename}_Center" "${data_center[*]}"
-	for i in "${data_center[@]}"
-	do
-	    echo "Combining run $i with ${OutDATAFilename}_Center.root..."  
-	    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
-	    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
-	done
+	if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${OutDATAFilename}_Center.root" ]; then
+	    for i in "${data_center[@]}"
+	    do
+		if [ -f "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT/${i}_-1_Raw_Data.root" ]; then
+		    cd "${LTANAPATH}/OUTPUT/Analysis/${ANATYPE}LT"
+		    echo "Combining run $i with ${OutDATAFilename}_Center.root..."
+		    echo "Renaming ${i}_Raw_Data to ${i}_Proc_Data..."
+		    #mv ${i}_-1_Raw_Data.root ${i}_-1_Proc_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		    mv ${i}_-1_Proc_Data.root ${i}_-1_Raw_Data.root # <runNum>_-1_Proc_Data.root is used in later LT_analysis
+		else
+		    echo "WARNING: ${i}_Raw_Data.root does not exist!"
+		fi
+	    done	 
+	fi
 	echo
-    fi
+    fi    
     
 fi
